Make static context cross-compiler safe.
Generated context is now an array of ge_storage_t rather than a byte array, to avoid endianness and packing issues.
A new m4 file and automake rules are used to make sure gen_context is built using the correct compiler.
A basic config is added for the gen_context tool to use.